Update api.py with proper base classes
Also, restructure `tests/` directory and remove unwanted file
